requirements.txt
anyio==3.6.1
argon2-cffi==21.1.0
asgiref==3.4.1
async-generator==1.10
attrs==21.2.0
backcall==0.2.0
bleach==4.1.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.8
cycler==0.11.0
debugpy==1.5.1
defusedxml==0.7.1
distlib==0.3.4
dj-database-url==0.5.0
Django==3.2.16
django-ckeditor==6.1.0
django-crispy-forms==1.12.0
django-js-asset==1.2.2
django-phonenumber-field==5.2.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
execnet==1.9.0
fastapi==0.79.0
filelock==3.4.2
fonttools==4.33.3
greenlet==1.1.2
h11==0.13.0
idna==3.3
importlib-resources==5.4.0
iniconfig==1.1.1
ipykernel==6.5.0
ipython==7.29.0
ipywidgets==7.6.5
jedi==0.18.0
Jinja2==3.0.2
jsonschema==4.2.1
jupyter==1.0.0
kiwisolver==1.4.2
MarkupSafe==2.0.1
mistune==0.8.4
nbclient==0.5.5
nbconvert==6.2.0
nbformat==5.1.3
nest-asyncio==1.5.1
notebook==6.4.5
numpy==1.22.4
openpyxl==3.0.7
outcome==1.1.0
packaging==21.0
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.2
pexpect==4.8.0
phonenumbers==8.12.29
pickleshare==0.7.5
Pillow==9.0.1
platformdirs==2.4.1
pluggy==0.13.1
prometheus-client==0.12.0
prompt-toolkit==3.0.22
psycopg2==2.9.1
ptyprocess==0.7.0
py==1.10.0
pycparser==2.21
pydantic==1.9.1
Pygments==2.10.0
pyOpenSSL==22.0.0
pyparsing==2.4.7
pyrsistent==0.18.0
PySocks==1.7.1
pytest==6.2.4
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-xdist==2.4.0
python-dateutil==2.8.2
pytz==2021.1
pyzmq==22.3.0
qtconsole==5.2.0
QtPy==1.11.2
requests==2.26.0
selenium==3.141.0
Send2Trash==1.8.0
sniffio==1.2.0
sortedcontainers==2.4.0
SQLAlchemy==1.4.39
sqlparse==0.4.2
starlette==0.19.1
terminado==0.12.1
testpath==0.5.0
toml==0.10.2
tomli==2.0.1
tornado==6.1
trio==0.20.0
trio-websocket==0.9.2
typing_extensions==4.3.0
urllib3==1.26.6
uvicorn==0.18.2
virtualenv==20.13.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==5.3.0
widgetsnbextension==3.5.2
wsproto==1.1.0
zipp==3.6.0
